requirements.txt
runpod>=1.7.0
requests>=2.28.0
Pillow>=10.0.0
numpy>=1.24.0
rembg>=2.0.50
onnxruntime-gpu>=1.16.0
spandrel>=0.1.0
torch>=2.0.0
opencv-python-headless>=4.8.0
